Support show statements based on AST editing
This is a major rework of the show statement support. For a clean
processing, the AST of the ShowTerms is analzyed and edited in
2 steps:
- First, the ShowTerms are analyzed by the ProgramAnalyzer similar
  to how rules are.
- Second, ShowTerms are transformed into a rule in the justifier Program
  of the form
  `h(I, (showTerm.term), (reasons)) :- showTerm(showTerm.term),
                                       showTerm.body.`
  This rule differs from the standard justifier rule in that it has the
  tuple of the showTerm.term in the head, but that tuple is wrapped in
  the `showTerm` literal in the body. This is needed, because the tuple
  allows language constructs that are not allowed in the body of a rule.
The second step proved somewhat tricky, as we need a different term in
the wrapped stable model (in the facts of the justifier program) than
it appears in the stable model. To this end, all terms in the stable
model are wrapped in this `showTerm` literal. This is needed, as there
may be touples in the "head" of a showTerm that are not allowed as
facts. To find a conflict free name of the `showTerm` literal, an
entirely new name tracker was implemented in the ProgramAnalyzer.
During the Program analysis step, tracker collects all strings used in
the program from which conflicts could arise. Later, this list is used
to find conflict free names of `h`, `model`, variables, and `showTerm`.

Resolves: potassco/viasp#23