Add fix to always show showTerm terms in the graph
The transformation of showTerms into justifier rules in the justifier
program with the `h` literal in the head automatically went through
some post-processing. Specifically the part which prevents atoms to
show up multiple times in the graph, if it could be derived by multiple
rules. To skip this post-processing, the `h` literal for showTerms is
renamed to `h_showTerm`. After the justifier program is grounded,
the association of the `h_showTerm` literal of number to symbol
is never skipped due to preexisting atoms in the graph.

Resolves: potassco/viasp#57