Improve backend type hinting
This commit adds improved type hints to the backend codebase. It
especially adds the types for old code. The new code is already typed,
but was significantly improved as well. The type hints are now more
precise and more consistent. For AST Transformers, the typing focuses
on readability and not on correctness. This is because the AST type are
not accepted by type checkers.

Resolves: potassco/viasp#60