Introduce draggable rules to viASP frontend
This commit introduces additional fixes to the viASP frontend in the
context of the draggable rules feature, while this message summarizes
previous commits.
In order to store multiple possible graph orders, the backend now stores
all graphs in a database. Each order has a unique hash, the composite
hash of the transformations' hashes. The frontend now requests the
hashes of all sorts from the backend and sets the first one as the
current sort. It then requests the transformations and nodes of the
current sort.
The rows are placed in a DraggableList container.
Upon dragging a row, the frontend calculates where the row can be
dropped by comparing composite hashes of all dropping positions
(calcualted by the browser) to the previously stored possible sorts'
hashes. Dropping positions that are not allowed are indicated by graying
out the row.
Upon dropping a row in a new position, highlighted symbols and shown
recursion are reset, and the frontend posts the new sort to the backend,
which sets this as the new default graph. The frontend then reloads the
transformations and nodes. All UUIDs in the graph change in this
process.
To indicate to the user the reloading of the graph a new loading screen
is introduced. Normal nodes are replaced with 'loading nodes', which
have no content and an animation. Once all transformations and nodes
are fully loaded, the loading nodes are replaced (all at the same time).
Due to these changes, API calls from the frontend have been improved
and refactored. Most of them happen in the corresponding contexts
instead of the components.

Contributes: potassco/viasp#24