Add clingraph behavior for dragged rules
A better way of handling clingraph visualizations is provided by
streamlining clingraph with the graph data. The frontend now manages
both clingraph data (links to files) and graph data (nodes and edges) in
the transformations context.
The backend stores clingraph file names in the database.
This approach allows a similar shading behavior while the user is
dragging rows, and similar loading behavior upon reloading of the graph.

Contributes: potassco/viasp#24