Fix tests and CI/CD workflows
For consistent test results, teardown code was added to the pytest setup
of the flask app.
The CI/CD workflows were updated to use the latest versions of the
actions.
The path of the colorPalette.json file was changed and the file is now
included in the setup file of the backend.