Drop Aggregate Reasons, improve Anon Renaming
The inclusion of Aggregate elements and conditions of conditional
literals in the body of a rule, as the reasons of the head literal
is dropped in this version, becuase it lacks direct semantic meaning.
This simplifies the building of justifier rules, due to a simpler
collection of body elements.

The renaming of anonymous variables is improved. Negated literals
are not edited.