Add minimization support and efficient graph creat
This commit adds minimization support to the viASP cli and the python
API. The available options are equivalent to clingo. They are passed on
to the clingo control object to be tasked with optimization.

The graph creation has been optimized to be more efficient. Programs
with many rules and correspondingly many possible sorts are now created
on demand. This is done by generating only the graph from the primary
sort in the beginning. All possible sorts are calculated and stored in
the database. When the frontend requests a graph for one of the possible
sorts, it is created while the frontend shows a loading animation.

All data is stored in the sqlite database. It is created and destroyed
when the server is started and stopped.

Resolves: potassco/viasp#34