Add simpler Dependency Collection
This commit includes a new DependencyCollector class that is used to
parse the heads and bodies of ASP rules and rule-like constructs for
the ProgramAnalyzer and the ProgramReifier.

When heads are parsed, Conditional Literals need to separate the
dependants from conditions.

when bodies are parsed, Literals are Collected as conditions of the
rule. Only when parsed for the analyzer, all Literals (except for
aggregates) are collected as conditions. For the reifier, only Literals
that are not located in aggregatese are collected as conditions.
Conditional Literals in the body are also collected as conditions.

To keep the code clean, the Analyzer's methods for visiting Rules,
ShowTerms, and Minimize have been consolidated.