Improve generation of recursive sub-graphs
The main change in this commit is the clarification of how viasp deals
with recursive sub-graphs of rules that contain aggregates. This is
described in issue potassco/viasp#62.

Refactoring of the justifier program creation. It uses the transformer
architecture to generate justifier rules from more comples rules, e.g.
ones containing aggregates. The wrapping in `model` literals is also
managed in a new transformer.

Fixes to the analyzer and reifier to support collection of positive
conditions. They are now collected in the transformer's visiting
process. This way, we know the context of each literal (e.g. in
aggregate, negated, etc.).

Contributes: potassco/viasp#62