Add python API endpoint for quick startup
This commit adds a new endpoint to the viasp Python API that allows for
a quick startup of viasp visualizations. This can be used as an
equivalent to the command line interface, for example in Jupyter
notebooks. Note that in binder, this does not work properly, because
`jupyter_dash.infer_jupyter_proxy_config()` needs to be run in a separate
cell.

The timeout duration for server startup is now managed with the retrying
package. This also shortens the time waited for failed startups.

Contributes: potassco/viasp#42
Resolves: potassco/viasp#69
Resolves: potassco/viasp#68