Add timeout to sortable feature
This commit adds a timeout to the sortable feature in the Row component.
If the backend takes too long to generate all possible sorts, only
the primary sort (as dictated by the order of the rules in the input
program) is shown. This will prevent long loading times.

Additionally, the drag handle is omitted in the frontend if the
sort generation timed out or if there is only one possible sort.

some numbers: say it takes 1s to generate and save 1000 sorts. A patient
user will wait 30s for the program to do anything. That's 3e^4 sorts.
That is reached by 8 independent rules.