Introduce lazy loading of possible orders
This approach prevents a huge computational load at the beginning of
the rendering process. Possible orders are calculated 'lazily', i.e.
when needed.

Upon the initial rendering of the graph, each component (transformation)
is analyzed in the current (primary) sort order. If the component can
be placed at a different index with all other transformations in the
same order, this new order is called "adjacent" to the current order.

This index is stored in a component's adjacent_sort_indices attribute.
Additionally the hash of this new order is generated and stored in a
the graphs database table. A database table "graph_relations" stores
this adjacency relation between the (current) graph's hash and the
adjacent graph's hash.

The frontend doesn't fetch possible orders directly, instead it follows
the possible orders defined in the component's attributes. When the user
drags a component to a new index, the new index is compared to the list
of possible indices in the component's adjacent_sort_indices attribute.

If the new index is in the list of possible indices, the component is
repositioned and the information about the moved component is sent to
the backend. The backend then calculates the adjacent_sort_indices for
the new order.

Contributes: potassco/viasp#34