Use RuleContainer to manage rules
The rule container can be stored in the database, so that the input
program does never need to be parsed again. The Container has both the
AST representation as well as the string representation of the rules.

When the dependency graph is generated, the RuleContainer is created
from the input AST. A rule container can then be merged with another
to remove cycles/loops in the graph. A component of the viasp graph
then contains a RuleContainer.

This commit contains changes to the whole codebase to reflect the new
Container. Tests are updated.

Contributes: potassco/viasp#34