Improve recursion justifier program
This commit improves the recursion justifier program for showing
correct recursive subgraphs for recursion with choice.

For example, the rule for a loop recursion with choice is as follows:
```
% input rule:
{b(X)} :- b(X-1), a(X).

% justifier rule:
h(n,b(X),(b((X-1)),a(X))) :-
	model(a(X));
	model(b((X-1)));
	@derivable(b(X)) = 1,
	not model(b(X)).
```

The addition is the @derivable(b(X)) = 1 condition, which is used to
check if the super-node contains the b(X) symbol. If it is in the super-
node, it can be derived. If it is not in the super-node, it cannot be
derived.

Resolves: potassco/viasp#12