Preparation commit for expanding Row
This commit is a preparation commit for expanding Rows instead of nodes.
Clicking on a row will expand all nodes in that row.

This commit moves the overflow information to the transformations
context as part of the nodes stored in there. This is necessary to
make the values available to the Row component, but also make it
possible for individual nodes to set their own values using the
visisbility Manager.