Expand rows, if node is folded
Modifying the way in which the node expansion state is stored. This
should improve performance.

The recrusive subnodes are not stored as a graph, as they are always
just a list of nodes chained in one path (order determined by the
recursion index). (See potassco/viasp#12)

The button in the bottom of the row is now fully functional.