Add generalized signature function to analyzer
The analyzer passes AST nodes to the make signature function in order to
generate a signature for the node in the dependency graph.

This commit adds a generalized signature function that is used to
generate signatures for all possible AST nodes. `None` is returned for
AST that is irrelevant for the dependency graph (e.g. aggregates).

Adding tests for all possible AST nodes that can be analyzed.